docs: split Python stub module
