+++ src/lib.rs
@@ -12,6 +12,8 @@ mod typing;
 use async_impl::{Client, Message, Response, Streamer, WebSocket};
 use blocking::{BlockingClient, BlockingResponse, BlockingStreamer, BlockingWebSocket};
 use error::*;
+use pyo3::types::PyDict;
+use pyo3::wrap_pymodule;
 use pyo3::{prelude::*, pybacked::PyBackedStr};
 use pyo3_async_runtimes::tokio::future_into_py;
 use typing::param::{RequestParams, WebSocketParams};
@@ -140,40 +142,107 @@ fn websocket(
 }
 
 #[pymodule(gil_used = false)]
-fn rnet(m: &Bound<'_, PyModule>) -> PyResult<()> {
+fn rnet(py: Python, m: &Bound<'_, PyModule>) -> PyResult<()> {
     pyo3::prepare_freethreaded_python();
-
-    m.add_class::<Method>()?;
-    m.add_class::<Version>()?;
-    m.add_class::<HeaderMap>()?;
-    m.add_class::<HeaderMapItemsIter>()?;
-    m.add_class::<HeaderMapKeysIter>()?;
-    m.add_class::<HeaderMapValuesIter>()?;
-    m.add_class::<Impersonate>()?;
-    m.add_class::<ImpersonateOS>()?;
-    m.add_class::<ImpersonateOption>()?;
-    m.add_class::<TlsVersion>()?;
     m.add_class::<SocketAddr>()?;
-    m.add_class::<Proxy>()?;
-    m.add_class::<LookupIpStrategy>()?;
     m.add_class::<Message>()?;
     m.add_class::<StatusCode>()?;
     m.add_class::<Part>()?;
     m.add_class::<Multipart>()?;
-
-    m.add_class::<SameSite>()?;
-    m.add_class::<Cookie>()?;
-
     m.add_class::<Client>()?;
     m.add_class::<Response>()?;
     m.add_class::<WebSocket>()?;
     m.add_class::<Streamer>()?;
+    m.add_class::<Proxy>()?;
+
+    m.add_function(wrap_pyfunction!(get, m)?)?;
+    m.add_function(wrap_pyfunction!(post, m)?)?;
+    m.add_function(wrap_pyfunction!(put, m)?)?;
+    m.add_function(wrap_pyfunction!(patch, m)?)?;
+    m.add_function(wrap_pyfunction!(delete, m)?)?;
+    m.add_function(wrap_pyfunction!(head, m)?)?;
+    m.add_function(wrap_pyfunction!(options, m)?)?;
+    m.add_function(wrap_pyfunction!(trace, m)?)?;
+    m.add_function(wrap_pyfunction!(request, m)?)?;
+    m.add_function(wrap_pyfunction!(websocket, m)?)?;
+
+    // Header module
+    m.add_wrapped(wrap_pymodule!(header_module))?;
+
+        // Cookie module
+    m.add_wrapped(wrap_pymodule!(cookie_module))?;
+
+    // Impersonate module
+    m.add_wrapped(wrap_pymodule!(impersonate_module))?;
+
+    // Enum module
+    m.add_wrapped(wrap_pymodule!(enums_module))?;
+
+    // Blocking module
+    m.add_wrapped(wrap_pymodule!(blocking_module))?;
+
+    // Exceptions module
+    m.add_wrapped(wrap_pymodule!(exceptions_module))?;
+
+    // Inserting to sys.modules allows importing submodules nicely from Python
+    // e.g. from maturin_starter.submodule import SubmoduleClass
+
+    let sys = PyModule::import(py, "sys")?;
+    let sys_modules: Bound<'_, PyDict> = sys.getattr("modules")?.downcast_into()?;
+    sys_modules.set_item("rnet.header", m.getattr("header")?)?;
+    sys_modules.set_item("rnet.cookie", m.getattr("cookie")?)?;
+    sys_modules.set_item("rnet.impersonate", m.getattr("impersonate")?)?;
+    sys_modules.set_item("rnet.enums", m.getattr("enums")?)?;
+    sys_modules.set_item("rnet.blocking", m.getattr("blocking")?)?;
+    sys_modules.set_item("rnet.exceptions", m.getattr("exceptions")?)?;
+    Ok(())
+}
+
+#[pymodule(name = "header")]
+fn header_module(m: &Bound<'_, PyModule>) -> PyResult<()> {
+    m.add_class::<HeaderMap>()?;
+    m.add_class::<HeaderMapItemsIter>()?;
+    m.add_class::<HeaderMapKeysIter>()?;
+    m.add_class::<HeaderMapValuesIter>()?;
+    Ok(())
+}
+
+#[pymodule(name = "cookie")]
+fn cookie_module(m: &Bound<'_, PyModule>) -> PyResult<()> {
+    m.add_class::<Cookie>()?;
+    Ok(())
+}
+
+#[pymodule(name = "impersonate")]
+fn impersonate_module(m: &Bound<'_, PyModule>) -> PyResult<()> {
+    m.add_class::<Impersonate>()?;
+    m.add_class::<ImpersonateOS>()?;
+    m.add_class::<ImpersonateOption>()?;
+    Ok(())
+}
+
+#[pymodule(name = "enums")]
+fn enums_module(m: &Bound<'_, PyModule>) -> PyResult<()> {
+    m.add_class::<Method>()?;
+    m.add_class::<Version>()?;
+    m.add_class::<LookupIpStrategy>()?;
+    m.add_class::<TlsVersion>()?;
+    m.add_class::<SameSite>()?;
+    Ok(())
+}
+
+#[pymodule(name = "blocking")]
+fn blocking_module(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<BlockingClient>()?;
     m.add_class::<BlockingResponse>()?;
     m.add_class::<BlockingWebSocket>()?;
     m.add_class::<BlockingStreamer>()?;
+    Ok(())
+}
 
-    let py = m.py();
+
+#[pymodule(name = "exceptions")]
+fn exceptions_module(py: Python, m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add("DNSResolverError", py.get_type::<DNSResolverError>())?;
     m.add("BodyError", py.get_type::<BodyError>())?;
     m.add("BuilderError", py.get_type::<BuilderError>())?;
@@ -190,17 +259,5 @@ fn rnet(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add("UpgradeError", py.get_type::<UpgradeError>())?;
     m.add("URLParseError", py.get_type::<URLParseError>())?;
     m.add("MIMEParseError", py.get_type::<MIMEParseError>())?;
-
-    m.add_function(wrap_pyfunction!(get, m)?)?;
-    m.add_function(wrap_pyfunction!(post, m)?)?;
-    m.add_function(wrap_pyfunction!(put, m)?)?;
-    m.add_function(wrap_pyfunction!(patch, m)?)?;
-    m.add_function(wrap_pyfunction!(delete, m)?)?;
-    m.add_function(wrap_pyfunction!(head, m)?)?;
-    m.add_function(wrap_pyfunction!(options, m)?)?;
-    m.add_function(wrap_pyfunction!(trace, m)?)?;
-    m.add_function(wrap_pyfunction!(request, m)?)?;
-    m.add_function(wrap_pyfunction!(websocket, m)?)?;
-
     Ok(())
 }
